+++ Progetto OO/src/Classi/Confezionato.java
@@ -6,8 +6,8 @@ public class Confezionato extends Prodotto {
 	private String TipoConfezione;
 	
 	//Costruttore
-	public Confezionato(String nome, String paeseDiProvenienza, double quantitaNegozio, double prezzoUnitario,
-			int scontoPercentuale, double quantitaDeposito, Classi.Fornitura fornitura, String tipoConfezione) {
+	public Confezionato(String nome, String paeseDiProvenienza, float quantitaNegozio, float prezzoUnitario,
+			int scontoPercentuale, float quantitaDeposito, Classi.Fornitura fornitura, String tipoConfezione) {
 		super(nome, paeseDiProvenienza, quantitaNegozio, prezzoUnitario, scontoPercentuale, quantitaDeposito,
 				fornitura);
 		TipoConfezione = tipoConfezione;
