+++ Progetto OO/src/Classi/Verdura.java
@@ -8,8 +8,8 @@ public class Verdura extends Prodotto {
 	private LocalDate DataRaccolta;
 	
 	//Costruttore
-	public Verdura(String nome, String paeseDiProvenienza, double quantitaNegozio, double prezzoUnitario,
-			int scontoPercentuale, double quantitaDeposito, Classi.Fornitura fornitura, LocalDate dataRaccolta) {
+	public Verdura(String nome, String paeseDiProvenienza, float quantitaNegozio, float prezzoUnitario,
+			int scontoPercentuale, float quantitaDeposito, Classi.Fornitura fornitura, LocalDate dataRaccolta) {
 		super(nome, paeseDiProvenienza, quantitaNegozio, prezzoUnitario, scontoPercentuale, quantitaDeposito,
 				fornitura);
 		DataRaccolta = dataRaccolta;
